feat(refs T31492): Move boilerplate to statement folder
It is not only used for segments, so it is better placed in the statement folder.
