+++ client/js/components/procedure/StatementSegmentsList/StatementSegment.vue
@@ -270,7 +270,7 @@
 import { checkResponse, dpApi } from '@demos-europe/demosplan-utils'
 import { CleanHtml, DpButtonRow, DpCheckbox, DpIcon, DpLabel, DpMultiselect, prefixClassMixin, VPopover } from '@demos-europe/demosplan-ui'
 import { mapActions, mapMutations, mapState } from 'vuex'
-import DpBoilerPlateModal from '@DpJs/components/statement/segments/DpBoilerPlateModal'
+import DpBoilerPlateModal from '@DpJs/components/statement/DpBoilerPlateModal'
 import DpClaim from '@DpJs/components/statement/DpClaim'
 import DpRecommendationModal from '@DpJs/components/statement/segments/DpRecommendationModal'
