+++ client/js/components/statement/DpBoilerPlateModal.vue
@@ -42,7 +42,7 @@
 <script>
 import { DpButtonRow, DpModal } from '@demos-europe/demosplan-ui'
 import { mapActions, mapGetters, mapState } from 'vuex'
-import DpBoilerPlate from '@DpJs/components/statement/segments/DpBoilerPlate'
+import DpBoilerPlate from '@DpJs/components/statement/DpBoilerPlate'
 
 export default {
   name: 'DpBoilerPlateModal',
