+++ client/js/components/statement/assessmentTable/DpBulkEditStatement.vue
@@ -213,7 +213,7 @@
 import { checkResponse, dpApi } from '@demos-europe/demosplan-utils'
 import { DpButton, DpMultiselect, DpTextWrapper, prefixClassMixin } from '@demos-europe/demosplan-ui'
 import { mapActions, mapGetters, mapState } from 'vuex'
-import DpBoilerPlateModal from '@DpJs/components/statement/segments/DpBoilerPlateModal'
+import DpBoilerPlateModal from '@DpJs/components/statement/DpBoilerPlateModal'
 import { v4 as uuid } from 'uuid'
 
 export default {
